database/requirements.txt
mysql-connector-python==8.1.0
Flask-Bcrypt==1.0.1
pyotp==2.9.0
